Big refactor of code: now the loss functions are organized into a package with multiple modules.
