+++ u_rpn/data_preparation/generate_file.py
@@ -65,7 +65,7 @@ augmentation = [  # apply the following augmenters to most images
 augmentation = iaa.Sequential(augmentation)
 
 TO_GENERATE = 300
-OUTPUT_FOLDER = os.path.join(".", "out", "augmented")
+OUTPUT_FOLDER = os.path.join("..", "..", "out", "augmented")
 
 
 def get_contour_precedence(contour, cols):
@@ -75,7 +75,7 @@ def get_contour_precedence(contour, cols):
 
 
 def main():
-    path_regions = os.path.join(".", "in", "train")
+    path_regions = os.path.join("..", "..", "in", "train")
     images_info_path = os.path.join(path_regions, "via_region_data.json")
     # images_info = list(images_info.values())  # We do not need the keys
     images_info = normalize_dataset.get_raw_img_and_info(path_regions, images_info_path)
