+++ u_rpn/data_preparation/normalize_dataset.py
@@ -12,24 +12,24 @@ import warnings
 import numpy as np
 import cv2
 
-import common_main_funcs as cmf
+from u_rpn.common import dataset_funcs as dtf
 
 warnings.filterwarnings("ignore")
 
-OUTPUT_FOLDER = os.path.join(".", "out", "normalized")
+OUTPUT_FOLDER = os.path.join("../..", "out", "normalized")
 
 
 def main():
-    path_regions = os.path.join(".", "in", "train")
+    path_regions = os.path.join("../..", "in", "train")
     images_info_path = os.path.join(path_regions, "via_region_data.json")
 
     os.makedirs(OUTPUT_FOLDER, exist_ok=True)
 
     bboxes_json = {}
 
-    for img_idx, img, regions, mask, cell_type in cmf.get_raw_img_and_info(path_regions,
+    for img_idx, img, regions, mask, cell_type in dtf.get_raw_img_and_info(path_regions,
                                                                            images_info_path):
-        img, improved_mask, regions_augmented = cmf.normalize_img_mask(img, mask)
+        img, improved_mask, regions_augmented = dtf.normalize_img_mask(img, mask)
 
         img_idx = str(img_idx).zfill(2)
         bboxes_json[img_idx] = {'regions': regions_augmented, 'cell_class': cell_type}
